Determine the name of the resource bundle at compile time
Since we already need a file that lists all available locales, we might as
well put the name of our resource bundle into it, too. To keep things sane,
we write a clojure map into resources/locales.clj from the msgfmt target
and read that to get at the bundle name and locales list.