Support mutliple locales.clj files on classpath
For projects that are use libraries that themselves contain resource
bundles, make it possible to effectively combine all the message catalogs
of each individual project into one big (logical) message catalog

We actually do not combine message catalogs at all; instead, the tru and
trs macros pass the current namespace along, and we use that to look the
message up in the corresponding resource bundle.

The format of locales.clj has changed to make it possible to find the
resource bundle for a project based on the namespace. The assumption is
that all namespaces in a project are nested under the project name, and
that there are no two projects for which the name of the first is a prefix
of the name of the second.