Introduce the notion of a MESSAGE_LOCALE and generate a resource bundle for it
Our raw messages in messages.pot are in some locale, namely English. Add
that to the list of available locales, and make an explicit resource bundle
for it.

Expose that locale as (message-locale) in Clojure so that we have a clear
fallback for when we can't find a matching locale for the user. We can't
necessarily use the system-locale as that might be very far from what the
user can understand; English seems like the safest fallback in these
situations.

Using MESSAGE_LOCALE here gives the user the same behavior when we can't
find a matching locale as they get when a message isn't in the message
catalog, or hasn't been localized yet.

As a small side-benefit, we can use the corresponding resource bundle
explicitly when we can't set a user-locale according to the user's wishes,
rather than special casing "we don't have a user-locale" everywhere.