(INTL-2) Move generated files to `:target-path`
This commit moves the Makefile.i18n to `:target-path` and generates the
Makefile.i18n as part of the compile-hook for the plugin. This way we
ensure the Makefile never falls out of date with the plugin. This commit
also changes the Makefile.i18n so that our generated class files end up
in `target/classes` (leins `:compile-path`) as well.