(INTL-1) Support for multiple packages in locales.clj
Prior to this commit the :package key in the locales.clj file as generated
by the Makefile was set to a value derived from the project name.
This didn't work at least for the following projects:
 * puppetlabs/pcp-common
 * puppetlabs/pcp-client
which don't use the assumed prefixes of 'puppetlabs.pcp_common' and
'puppetlabs.pcp_client' respectively for their namespaces but rather define
different namespaces under the common root of 'puppetlabs.pcp'.
To be able to support that, this patch redefines the structure of the
locales.clj such that it can now contain the ':packages' key (as opposed
to the original ':package') which is expected to contain a collection
of packages covered by the associated ':bundle'.

The Makefile is updated to support this new structure. There are couple of
user impacting changes:
 * the Makefile's PACKAGE variable is renamed to BUNDLE
 * a new variable named PACKAGES (note the plural) is introduced; it defines
   the list of packages supported by the resource bundle; by default it is
   set to $(BUNDLE) to mimic the original behavior, but it can be overridden
   by the user should the project need it