(INTL-2) Refactored i18n-make to a separate ns
This commit moves most of the logic out of the `leiningen.i18n` plugin
and into a separate namespace that can be invoked from a test
namespace. What's left in `leiningen.i18n` is just the stuff needed to
make it a task (i.e. help text, dealing with the project etc). This
commit also removes the need to out-of-band run make before running the
tests.