(ORCH-1232) Add a standard function for marking strings for extraction
Previously, the only way to mark a string for extraction was to either
use a translation function or use a project-specific marker function and
update the make target accordingly. This commit provides a standard
i18n/mark function which has no behavior beyond returning its argument,
but which serves as a designated way to extract strings without
translating them.

This is important for cases where the string needs to be translated
dynamically but is defined at compile time. In particular, it allows for
commonly-used strings to be defined in a variables and used in multiple
places.