(PE-19745) Empty string should render an empty string not POT header.
Prior to this commit, if a string was empty, instead of an empty string in the class bundle, GNU gettext behavior is to print the header of the POT file. So where we would expect empty strings, the POT header would be printed instead.

With this commit if a msgid is an empty string, an empty string is returned.