(maint) Don't create an English resource bundle.
The creation of the English resource bundle Messages_en.class was
causing the POT file to still be updated via the default msgfmt target
that's invoked from the leiningen compile hook. This results in the POT
file still being automatically updated during development, which we
don't want now that there's a CI job that handles updating the POT file.

Rectify this by removing the en MESSAGE_LOCALE from all the locale
lists and removing the target to generate the Messages_en.class. Since
the library returns the original message when there's no bundle for
a locale, and all our messages are in the English locale already, this
should not cause a change in behavior.

Since this introduces the possibility of having an orphaned English
resource bundle that's never updated (because there's no locales/en.po),
add a phony target called `clean-orphaned-bundles` that removes all
resource bundles that don't have a corresponding PO file, and make it
a dependency of the `msgfmt` target (so it will always run).