(PE-19944) Change structure of locales.clj parsing to accomodate lists
Prior to this commit the structure of the locales.clj was always assumed
to be a map. This commit changes the locales.clj parsing code to allow
the locales.clj to either a map or a list of maps.
This allows us to construct a list of the locales.clj maps when building
uberjars for multiple projects.