(PDB-5190) Run *all* system-locale fallback tests; fix related code
eastwood noticed that some tests weren't being run because they were
being driven by (lazy) map.  Replace map with doseq and then fix the
problems that revealed.

In particular, one of the test values was nil, and the existing code
couldn't handle that at all.  Attempt to adjust it to behave in what
might be the originally intended manner.