(PDB-5190) travis: check eastwood and reflection
eastwood found a number of issues, and quite a few more in puppetdb
when we added it there, so let's try requiring it for now.

And since there's no telling where a library's functions will be
used (e.g. in a tight loop), require all reflection warnings to be
fixed too.

The original impetus for this was the fact that (some very simplistic)
PuppetDB query performance profiling revealed that reflection in some
of the calls in i18n (e.g. underneath bundle-for-namespace, likely via
trs/tru) was showing up in the stack samples far more often for the
relevant threads than the functions that handle the actual query data.

If this requirement proves undesirable, we can always drop the it, or
just opt out of the warnings in relevant sections of code, perhaps
with something like this:

  (set! orig-warn-on-reflection *warn-on-reflection*)

  ...

  (set! orig-warn-on-reflection false)
  ...some code that opts out...
  (set! *warn-on-reflection* orig-warn-on-reflection)

  ...

  (set! orig-warn-on-reflection false)
  ...some more code that opts out...
  (set! *warn-on-reflection* orig-warn-on-reflection)